+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
 ## 1.5.21 (2020-06-01)
-
 - [713](https://github.com/technekes/cast-ui/pull/713): Adding prop for adjusting vertical spacing on DatePicker - [@MrMuzik](https://github.com/MrMuzik)
+- [708](https://github.com/technekes/cast-ui/pull/708): Pagination
+  enhancements with breaking changes: [@andrey-liventsev-ontarget](https://github.com/andrey-liventsev-ontarget)
+  - Pagination property btnSize was removed
+  - Pagination properties previousText/nextText have been replaced with PageButtonFirstLastComponent property
+  - rowsText default value was removed
 
 ## 1.5.20 (2020-05-27)